app-misc/powerline: Unit tests fixed.

+++ app-misc/powerline/powerline-9999-r20.ebuild
@@ -34,7 +34,7 @@ DEPEND="
 	doc? ( dev-python/sphinx[${PYTHON_USEDEP}] )
 	man? ( dev-python/sphinx[${PYTHON_USEDEP}] )
 	test? (
-		app-misc/screen
+		dev-python/pexpect
 		dev-python/psutil[${PYTHON_USEDEP}]
 		x11-libs/libvterm
 		>=dev-vcs/git-1.7.2
@@ -60,7 +60,8 @@ RDEPEND="
 "
 
 # Source directory from which all applicable files will be installed.
-POWERLINE_SRC_DIR="${T}/bindings"
+POWERLINE_BINDINGS_DIR="${T}/bindings"
+POWERLINE_FULL_BINDINGS_DIR="${T}/bindings-full"
 
 # Target directory to which all applicable files will be installed.
 POWERLINE_TRG_DIR='/usr/share/powerline'
@@ -129,7 +130,10 @@ python_prepare_all() {
 	# safely remove such files *AND* permit their installation after the main
 	# distutils-based installation, copy them to such location and then remove
 	# them from the original tree that distutils operates on.
-	cp -R "${S}"/powerline/bindings "${POWERLINE_SRC_DIR}" || die '"cp" failed.'
+	cp -R "${S}"/powerline/bindings "${POWERLINE_BINDINGS_DIR}" || die '"cp" failed.'
+	if use test ; then
+		cp -R "${S}"/powerline/bindings "${POWERLINE_FULL_BINDINGS_DIR}" || die '"cp" failed.'
+	fi
 
 	# Remove all non-Python files from the original tree.
 	find "${S}"/powerline/bindings -type f -not -name '*.py' -delete
@@ -139,7 +143,7 @@ python_prepare_all() {
 	# following unrelated Python files:
 	#
 	# * "powerline-awesome.py", an awesome-specific integration script.
-	find "${POWERLINE_SRC_DIR}"\
+	find "${POWERLINE_BINDINGS_DIR}"\
 		-type f\
 		-name '*.py'\
 		-not -name 'powerline-awesome.py'\
@@ -167,7 +171,20 @@ python_compile_all() {
 }
 
 python_test() {
-	PYTHON="${PYTHON}" "${S}"/tests/test.sh || die 'Unit tests failed.'
+	# *All* bindings files are required for tests.
+	mv "${S}"/powerline/bindings{,.bak}
+	cp -R "${POWERLINE_FULL_BINDINGS_DIR}" "${S}"/powerline/bindings
+	# Powerline shell tests do not work with LD_PRELOAD-based sandbox.
+	env -i \
+		USER="$USER" \
+		HOME="$HOME" \
+		LANG=en_US.UTF-8 \
+		PATH="$PATH" \
+		PYTHON="${PYTHON}" \
+		"${S}"/tests/test.sh \
+		|| die 'Unit tests failed.'
+	rm -r "${S}"/powerline/bindings
+	mv "${S}"/powerline/bindings{.bak,}
 }
 
 python_install_all() {
@@ -183,9 +200,9 @@ python_install_all() {
 	if use awesome; then
 		local AWESOME_LIB_DIR='/usr/share/awesome/lib/powerline'
 		insinto "${AWESOME_LIB_DIR}"
-		newins "${POWERLINE_SRC_DIR}"/awesome/powerline.lua init.lua
+		newins "${POWERLINE_BINDINGS_DIR}"/awesome/powerline.lua init.lua
 		exeinto "${AWESOME_LIB_DIR}"
-		doexe  "${POWERLINE_SRC_DIR}"/awesome/powerline-awesome.py
+		doexe  "${POWERLINE_BINDINGS_DIR}"/awesome/powerline-awesome.py
 
 		DOC_CONTENTS+="
 	To enable Powerline under awesome, add the following lines to \"~/.config/awesome/rc.lua\" (assuming you originally copied such file from \"/etc/xdg/awesome/rc.lua\"):\\n
@@ -195,7 +212,7 @@ python_install_all() {
 
 	if use bash; then
 		insinto "${POWERLINE_TRG_DIR}"/bash
-		doins   "${POWERLINE_SRC_DIR}"/bash/powerline.sh
+		doins   "${POWERLINE_BINDINGS_DIR}"/bash/powerline.sh
 
 		DOC_CONTENTS+="
 	To enable Powerline under bash, add the following line to either \"~/.bashrc\" or \"~/.profile\":\\n
@@ -204,7 +221,7 @@ python_install_all() {
 
 	if use busybox; then
 		insinto "${POWERLINE_TRG_DIR}"/busybox
-		doins   "${POWERLINE_SRC_DIR}"/shell/powerline.sh
+		doins   "${POWERLINE_BINDINGS_DIR}"/shell/powerline.sh
 
 		DOC_CONTENTS+="
 	To enable Powerline under interactive sessions of BusyBox's ash shell, interactively run the following command:\\n
@@ -213,7 +230,7 @@ python_install_all() {
 
 	if use dash; then
 		insinto "${POWERLINE_TRG_DIR}"/dash
-		doins   "${POWERLINE_SRC_DIR}"/shell/powerline.sh
+		doins   "${POWERLINE_BINDINGS_DIR}"/shell/powerline.sh
 
 		DOC_CONTENTS+="
 	To enable Powerline under dash, add the following line to the file referenced by environment variable \${ENV}:\\n
@@ -223,7 +240,7 @@ python_install_all() {
 
 	if use fish; then
 		insinto /usr/share/fish/functions
-		doins "${POWERLINE_SRC_DIR}"/fish/powerline-setup.fish
+		doins "${POWERLINE_BINDINGS_DIR}"/fish/powerline-setup.fish
 
 		DOC_CONTENTS+="
 	To enable Powerline under fish, add the following line to \"~/.config/fish/config.fish\":\\n
@@ -232,7 +249,7 @@ python_install_all() {
 
 	if use mksh; then
 		insinto "${POWERLINE_TRG_DIR}"/mksh
-		doins   "${POWERLINE_SRC_DIR}"/shell/powerline.sh
+		doins   "${POWERLINE_BINDINGS_DIR}"/shell/powerline.sh
 
 		DOC_CONTENTS+="
 	To enable Powerline under mksh, add the following line to \"~/.mkshrc\":\\n
@@ -263,7 +280,7 @@ python_install_all() {
 
 	if use rc; then
 		insinto "${POWERLINE_TRG_DIR}"/rc
-		doins   "${POWERLINE_SRC_DIR}"/rc/powerline.rc
+		doins   "${POWERLINE_BINDINGS_DIR}"/rc/powerline.rc
 
 		DOC_CONTENTS+="
 	To enable Powerline under rc shell, add the following line to \"~/.rcrc\":\\n
@@ -272,7 +289,7 @@ python_install_all() {
 
 	if use tmux; then
 		insinto "${POWERLINE_TRG_DIR}"/tmux
-		doins   "${POWERLINE_SRC_DIR}"/tmux/powerline*.conf
+		doins   "${POWERLINE_BINDINGS_DIR}"/tmux/powerline*.conf
 
 		DOC_CONTENTS+="
 	To enable Powerline under tmux, add the following line to \"~/.tmux.conf\":\\n
@@ -281,7 +298,7 @@ python_install_all() {
 
 	if use zsh; then
 		insinto /usr/share/zsh/site-contrib
-		doins "${POWERLINE_SRC_DIR}"/zsh/powerline.zsh
+		doins "${POWERLINE_BINDINGS_DIR}"/zsh/powerline.zsh
 
 		DOC_CONTENTS+="
 	To enable Powerline under zsh, add the following line to \"~/.zshrc\":\\n
